Preliminary implementation of operator overloading type checking
For some reason, if you build the project, the rules that are being
overridden to type check the usages of overloaded operators sometimes
defer to the core implementation. If you make trivial changes to the
checking rule (add a new line randomly) and rebuild, sometimes this
fixes itself and starts using the new rules. Perhaps this is an issue
order in which MPS is engaging the language generators?